project3/backend/00-data-ingestion/_archived/oanda-collector-backup/requirements.txt
# OANDA v20 API
v20==3.0.25

# Async HTTP client
aiohttp==3.9.1
asyncio==3.4.3

# NATS streaming
nats-py==2.6.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Monitoring
prometheus-client==0.19.0

# Protocol Buffers
protobuf==4.25.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
